Adding missing RPC - VEFBRPC
This RPC was not getting installed in the RAPTOR VistA instance, and was causing the EWD login call to fail. I installed it manually and the application works. This will need to be installed programmatically during the vagrant build process for the application to work right out of the box.