Added new info on KIDS install issue
KIDS package update for VistA does not install automatically. Installing it manually does not seem to work either. Installing the one routine - VEFBRPC - manually seems to work and allow login to succeed, but I'm not sure what the other routine - VEFBRAP0 - does and that isn't getting installed either so something else might not work in the app downstream if this isn't corrected.